Refactor RedisStore and MapDBStore
